requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-openai
langchain-ollama
langgraph
pydantic~=2.11.6
python-dotenv~=1.1.0